# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent System",
    "description": "AI20K Agent System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/sandbox/config": {
      "get": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Get Sandbox Config",
        "description": "Lấy cấu hình Sandbox DB của dự án.",
        "operationId": "getSandboxConfig",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án",
              "title": "Project Id"
            },
            "description": "ID dự án"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_SandboxConfigResponse__NoneType__"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Save Sandbox Config",
        "description": "Lưu hoặc cập nhật thông tin cấu hình Sandbox DB cho dự án.",
        "operationId": "saveSandboxConfig",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án",
              "title": "Project Id"
            },
            "description": "ID dự án"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SandboxConfigResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sandbox/test-connection": {
      "post": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Test Sandbox Connection Endpoint",
        "description": "Kiểm tra thử kết nối đến cơ sở dữ liệu Sandbox.",
        "operationId": "testSandboxConnection",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án",
              "title": "Project Id"
            },
            "description": "ID dự án"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TestConnectionResponse_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/sandbox/execute-ddl": {
      "post": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Execute Sandbox Ddl Endpoint",
        "description": "Thực thi mã DDL script trên Sandbox Database đã cấu hình của dự án.",
        "operationId": "executeSandboxDdl",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án",
              "title": "Project Id"
            },
            "description": "ID dự án"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteDdlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExecuteDdlResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Code",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông báo lỗi"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Mã lỗi ổn định cho client"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Danh sách chi tiết lỗi theo field"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "error_code"
        ],
        "title": "ApiErrorResponse",
        "description": "Error envelope được Global Exception Handler trả về."
      },
...
      "ApiResponse_ExecuteDdlResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecuteDdlResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[ExecuteDdlResponse]"
      },
...
      "ApiResponse_SandboxConfigResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxConfigResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[SandboxConfigResponse]"
      },
      "ApiResponse_TestConnectionResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestConnectionResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[TestConnectionResponse]"
      },
...
      "ApiResponse_Union_SandboxConfigResponse__NoneType__": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxConfigResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[Union[SandboxConfigResponse, NoneType]]"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Đường dẫn field không hợp lệ"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mô tả lỗi validation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Chi tiết validation của một trường request."
      },
      "ExecuteDdlRequest": {
        "properties": {
          "ddl_script": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Ddl Script",
            "description": "Nội dung mã DDL script cần chạy"
          },
          "reset_schema": {
            "type": "boolean",
            "title": "Reset Schema",
            "description": "Xoá và tạo lại schema đích trước khi chạy, trừ schema public.",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ddl_script"
        ],
        "title": "ExecuteDdlRequest",
        "description": "Payload thực thi DDL trên Sandbox."
      },
      "ExecuteDdlResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "executed_statements": {
            "type": "integer",
            "title": "Executed Statements"
          },
          "succeeded_statements": {
            "type": "integer",
            "title": "Succeeded Statements"
          },
          "failed_statements": {
            "type": "integer",
            "title": "Failed Statements"
          },
          "total_duration_ms": {
            "type": "number",
            "title": "Total Duration Ms"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/StatementLogResponse"
            },
            "type": "array",
            "title": "Logs",
            "description": "Log theo từng DDL statement"
          }
        },
        "type": "object",
        "required": [
          "success",
          "executed_statements",
          "succeeded_statements",
          "failed_statements",
          "total_duration_ms",
          "logs"
        ],
        "title": "ExecuteDdlResponse",
        "description": "Kết quả thực thi toàn bộ DDL script."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SandboxConfigRequest": {
        "properties": {
          "db_type": {
            "$ref": "#/components/schemas/SandboxDbType",
            "description": "Loại CSDL Sandbox",
            "default": "POSTGRESQL"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "description": "Host/IP kết nối",
            "default": "localhost"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "description": "Cổng kết nối DB",
            "default": 5432
          },
          "database_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Database Name",
            "description": "Tên CSDL",
            "default": "sandbox_db"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "Tên đăng nhập"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Mật khẩu kết nối"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1,
                "pattern": "^[A-Za-z_][A-Za-z0-9_]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name",
            "description": "Schema CSDL",
            "default": "public"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SandboxConfigRequest",
        "description": "Payload lưu cấu hình Sandbox."
      },
      "SandboxConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "db_type": {
            "$ref": "#/components/schemas/SandboxDbType"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "database_name": {
            "type": "string",
            "title": "Database Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name",
            "default": "public"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "CONFIGURED"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "db_type",
          "host",
          "port",
          "database_name"
        ],
        "title": "SandboxConfigResponse",
        "description": "Cấu hình Sandbox không làm lộ mật khẩu."
      },
      "SandboxDbType": {
        "type": "string",
        "enum": [
          "POSTGRESQL",
          "BIGQUERY",
          "SNOWFLAKE",
          "MYSQL",
          "SQLITE",
          "SQLSERVER"
        ],
        "title": "SandboxDbType",
        "description": "Loại cơ sở dữ liệu Sandbox."
      },
...
      "StatementLogResponse": {
        "properties": {
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "is_success": {
            "type": "boolean",
            "title": "Is Success"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          }
        },
        "type": "object",
        "required": [
          "statement",
          "is_success",
          "execution_time_ms",
          "timestamp"
        ],
        "title": "StatementLogResponse",
        "description": "Kết quả thực thi của một DDL statement."
      },
      "TestConnectionRequest": {
        "properties": {
          "db_type": {
            "$ref": "#/components/schemas/SandboxDbType",
            "description": "Loại CSDL Sandbox",
            "default": "POSTGRESQL"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host",
            "description": "Host/IP kết nối",
            "default": "localhost"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "description": "Cổng kết nối DB",
            "default": 5432
          },
          "database_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Database Name",
            "description": "Tên CSDL",
            "default": "sandbox_db"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "Tên đăng nhập"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Mật khẩu kết nối"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1,
                "pattern": "^[A-Za-z_][A-Za-z0-9_]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name",
            "description": "Schema CSDL",
            "default": "public"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TestConnectionRequest",
        "description": "Payload kiểm tra kết nối Sandbox."
      },
      "TestConnectionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "TestConnectionResponse",
        "description": "Kết quả kiểm tra kết nối Sandbox."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}